This borrows heavily from work that Ravago did to initially get this
stuff working.

Tested rudimentary functionality on a test bench, ensured that we could:
* Enable the robot.
* Read joystick and button values.
* Switch between auto and teleop modes.
* Read sensor values (encoder, absolute encoder, potentiometer).
* Read PDP values.
* Drive PWM motors.
* Drive CANivore motors.

Non-WPILib changes are made to accommodate the upgrade roborio
compiler's improved pickiness.

Merge commit '125aac16d9bf03c833ffa18de2f113a33758a4b8' into HEAD

+++ aos/network/web_proxy.h
@@ -229,7 +229,7 @@ class ApplicationConnection {
     bool requested = true;
   };
 
-  std::map<int, ChannelState> channels_;
+  std::map<size_t, ChannelState> channels_;
   const std::vector<std::unique_ptr<Subscriber>> &subscribers_;
 
   const std::vector<FlatbufferDetachedBuffer<MessageHeader>> config_headers_;
